requirements.txt
twilio==8.12.0
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.27.1
openai==1.12.0
requests==2.31.0
python-multipart==0.0.9 